+++ src/Mailer.php
@@ -136,8 +136,16 @@ class Mailer extends BaseMailer
             'Sender' => $message->getSender(),
             'Cc' => $message->getCc(),
             'Bcc' => $message->getBcc(),
+            'DeduplicateCampaign' => $message->getDeduplicateCampaign(),
             'TemplateErrorReporting' => $message->getTemplateErrorReporting() ? $message->getTemplateErrorReporting() : $this->getDefaultTemplateErrorReporting(),
         ];
+
+        if ($message->getCustomCampaign()){
+            $array['CustomCampaign'] = $message->getCustomCampaign();
+        }
+        if ($message->getCustomId()){
+            $array['CustomID'] = $message->getCustomId();
+        }
         
         $errorReporting = $message->getTemplateErrorReporting() ? $message->getTemplateErrorReporting() : $this->getDefaultTemplateErrorReporting();